tree command created
command prints the directory and its contents in a tree like format.
if no directory found, it prints the current directory, else it prints directory tree of all the directories provided through command line arguements.
At the end, it also shows the number of files and subfolders found in the directory.